+++ packages/terra-clinical-application-site/webpack.config.js
@@ -9,7 +9,8 @@ const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const I18nAggregatorPlugin = require('terra-i18n-plugin');
 const i18nSupportedLocales = require('terra-i18n/lib/i18nSupportedLocales');
-const CustomProperties = require('postcss-custom-properties');
+const PostCSSAssetsPlugin = require('postcss-assets-webpack-plugin');
+const PostCSSCustomProperties = require('postcss-custom-properties');
 const rtl = require('postcss-rtl');
 
 module.exports = {
@@ -49,7 +50,6 @@ module.exports = {
                       'iOS >= 8',
                     ],
                   }),
-                  CustomProperties(),
                   rtl(),
                 ];
               },
@@ -77,6 +77,13 @@ module.exports = {
     new I18nAggregatorPlugin({
       baseDirectory: __dirname,
       supportedLocales: i18nSupportedLocales,
+    }),
+    new PostCSSAssetsPlugin({
+      test: /\.css$/,
+      log: false,
+      plugins: [
+        PostCSSCustomProperties({ preserve: true }),
+      ],
     }),
     new webpack.NamedChunksPlugin(),
   ],